requirements.txt
fastapi
uvicorn[standard]
python-dotenv
openai
streamlit
requests
pydantic
pdfkit
Jinja2